GETP-216 rename: ClockHolder를 domain 모듈로 이동
