+++ src/main/java/es/princip/getp/domain/common/ClockHolder.java
@@ -1,4 +1,4 @@
-package es.princip.getp.common.domain;
+package es.princip.getp.domain.common;
 
 import java.time.Clock;
