+++ src/main/java/es/princip/getp/domain/project/apply/service/ProjectApplier.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.project.apply.service;
 
-import es.princip.getp.common.domain.ClockHolder;
+import es.princip.getp.domain.common.ClockHolder;
 import es.princip.getp.domain.common.model.AttachmentFile;
 import es.princip.getp.domain.common.model.Duration;
 import es.princip.getp.domain.people.exception.NotRegisteredPeopleProfileException;
