+++ src/main/java/es/princip/getp/domain/project/commission/service/ProjectCommissioner.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.project.commission.service;
 
-import es.princip.getp.common.domain.ClockHolder;
+import es.princip.getp.domain.common.ClockHolder;
 import es.princip.getp.domain.common.model.Duration;
 import es.princip.getp.domain.project.commission.exception.ApplicationDurationNotBeforeEstimatedDurationException;
 import es.princip.getp.domain.project.commission.exception.EndedApplicationDurationException;
