+++ src/test/java/es/princip/getp/domain/common/model/DurationTest.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.common.model;
 
-import es.princip.getp.common.infra.StubClockHolder;
+import es.princip.getp.domain.common.StubClockHolder;
 import es.princip.getp.domain.common.exception.StartDateIsAfterEndDateException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
