+++ src/test/java/es/princip/getp/domain/project/commission/model/ProjectStatusTest.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.project.commission.model;
 
-import es.princip.getp.common.infra.StubClockHolder;
+import es.princip.getp.domain.common.StubClockHolder;
 import es.princip.getp.domain.common.model.Duration;
 import org.junit.jupiter.api.Test;
